Included executables for the CLI version and made player search case-insensitve
